+++ R/methods-sampleYAML.R
@@ -4,7 +4,7 @@
 #' @name sampleYAML
 #' @family YAML Utilities
 #'
-#' @inheritParams AllGenerics
+#' @inheritParams general
 #'
 #' @param yaml Project summary YAML list.
 #' @param keys Nested operator keys, supplied as a character vector.
@@ -13,7 +13,7 @@
 #'   counts. Fast RNA-seq mode with lightweight counts (pseudocounts) doesn't
 #'   output the same metrics into the YAML.
 #'
-#' @return [data.frame].
+#' @return [tibble].
 #'
 #' @examples
 #' url <- file.path(
@@ -21,7 +21,7 @@
 #'     "bcbio",
 #'     "project-summary.yaml")
 #' yaml <- basejump::readYAML(url)
-#' sampleYAML(yaml, "metadata")
+#' sampleYAML(yaml, "metadata") %>% glimpse()
 NULL
 
 
@@ -31,43 +31,58 @@ NULL
 #' @importFrom dplyr arrange bind_rows
 #' @importFrom magrittr set_rownames
 .sampleYAML <- function(yaml, keys) {
-    samples <- yaml[["samples"]]
-    if (!length(samples)) {
-        abort("No sample information in YAML")
+    assert_is_non_empty(yaml)
+    yaml <- yaml[["samples"]]
+    assert_is_list(yaml)
+    assert_is_non_empty(yaml)
+
+    # Currently max 2 keys are supported
+    assert_all_are_in_range(length(keys), lower = 1L, upper = 2L)
+
+    # Always check that first key exists (e.g. "metadata")
+    assert_is_subset(keys[[1L]], names(yaml[[1L]]))
+
+    # When going deeper (e.g. metrics, check for secondary key presence)
+    if (length(keys) == 2L) {
+        assert_is_subset(
+            keys[[2L]],
+            names(yaml[[1L]][[keys[[1L]]]])
+        )
     }
-    if (!keys[[1L]] %in% names(samples[[1L]])) {
-        return(NULL)
-    }
-    if (length(keys) > 1L) {
-        if (!keys[[2L]] %in% names(samples[[1L]][[keys[[1L]]]])) {
-            return(NULL)
-        }
-    }
-    data <- lapply(samples, function(sample) {
-        nested <- sample[[keys]]
-        # Set the description
-        nested[["description"]] <- sample[["description"]]
+
+    data <- lapply(yaml, function(sample) {
+        # Always get the sample description
+        assert_is_character(sample[["description"]])
+        description <- sample[["description"]]
+
+        # Fix NULL batch and phenotype values found in metadata
         if (rev(keys)[[1L]] == "metadata") {
-            if (is.null(nested[["batch"]])) {
-                nested[["batch"]] <- NA
+            if (is.null(sample[[keys]][["batch"]])) {
+                sample[[keys]][["batch"]] <- NA
             }
-            if (length(nested[["phenotype"]])) {
-                if (grepl("^$", nested[["phenotype"]])) {
-                    nested[["phenotype"]] <- NA
+            if (length(sample[[keys]][["phenotype"]])) {
+                if (grepl("^$", sample[[keys]][["phenotype"]])) {
+                    sample[[keys]][["phenotype"]] <- NA
                 }
             }
         }
-        unlist <- unlist(nested)
-        names(unlist) <- camel(names(unlist), strict = FALSE)
+
+        unlist <- unlist(sample[[keys]])
+        names(unlist) <- camel(names(unlist))
+        # Add back the description
+        unlist <- c(description = description, unlist)
         unlist
     })
+
+    # Use this method to handle an uneven number of lengths.
+    # This can happen for sample metrics values that vary in lengths per sample.
     dflist <- lapply(data, function(x) {
-        as.data.frame(t(x), stringsAsFactors = FALSE)
+        as_tibble(t(x), stringsAsFactors = FALSE)
     })
+
     bind_rows(dflist) %>%
         removeNA() %>%
-        arrange(.data[["description"]]) %>%
-        set_rownames(.[["description"]])
+        arrange(.data[["description"]])
 }
 
 
@@ -77,6 +92,7 @@ NULL
 #' @export
 setMethod(
     "sampleYAML",
-    signature(yaml = "list",
-              keys = "character"),
+    signature(
+        yaml = "list",
+        keys = "character"),
     .sampleYAML)