+++ R/methods-sampleYAMLMetrics.R
@@ -6,8 +6,6 @@
 #'
 #' @inherit sampleYAML
 #'
-#' @inheritParams AllGenerics
-#'
 #' @examples
 #' url <- file.path(
 #'     "http://bcbiobase.seq.cloud",
@@ -22,22 +20,34 @@ NULL
 # Constructors =================================================================
 #' @importFrom dplyr mutate_if
 .sampleYAMLMetrics <- function(yaml) {
-    metrics <- sampleYAML(
+    # The fast mode RNA-seq pipeline doesn't report metrics generated from
+    # STAR featureCounts output with MultiQC
+    fastMode <- "Fast mode detected: No sample metrics were calculated"
+
+    # Early return on NULL metrics (fast mode)
+    if (is.null(yaml[["samples"]][[1L]][["summary"]][["metrics"]])) {
+        warn(fastMode)
+        return(NULL)
+    }
+
+    data <- sampleYAML(
         yaml = yaml,
         keys = c("summary", "metrics")
     )
-    # The fast mode RNA-seq pipeline doesn't report metrics generated from
-    # STAR featureCounts output with MultiQC. Allow NULL return to handle
-    # this pipeline output.
-    if (is.null(metrics)) {
-        warn("No sample metrics were calculated")
+    assert_is_tbl(data)
+
+
+    if (identical(colnames(data), "description")) {
+        warn(fastMode)
         return(NULL)
     }
+
     # Fix numerics set as characters
     numericAsCharacter <- function(x) {
         any(grepl(x = x, pattern = "^[0-9\\.]+$"))
     }
-    metrics %>%
+
+    data %>%
         mutate_if(is.factor, as.character) %>%
         mutate_if(numericAsCharacter, as.numeric) %>%
         mutate_if(is.character, as.factor) %>%