+++ R/uniteInterestingGroups.R
@@ -2,10 +2,9 @@
 #'
 #' Create a single interesting groups column (`interestingGroups`) used for
 #' coloring in plots. When multiple interesting groups are present, unite into a
-#' single column, delimited by ` : `.
+#' single column, delimited by `:`.
 #'
 #' @importFrom tidyr unite
-#' @importFrom rlang !!! syms
 #'
 #' @param object Object (e.g. [data.frame]) containing interesting groups
 #'   columns.
@@ -13,11 +12,29 @@
 #'
 #' @return [data.frame].
 #' @export
+#'
+#' @examples
+#' demultiplexed <- file.path(
+#'     "http://bcbiobase.seq.cloud",
+#'     "sample_metadata",
+#'     "demultiplexed.xlsx")
+#' meta <- readSampleMetadataFile(demultiplexed)
+#' meta <- uniteInterestingGroups(
+#'     object = meta,
+#'     interestingGroups = c("genotype", "sampleName")
+#' )
+#' pull(meta, "interestingGroups")
 uniteInterestingGroups <- function(object, interestingGroups) {
+    assert_has_colnames(object)
+    assert_is_character(interestingGroups)
+
     # Set up the interesting groups column
+    # TODO Move to an assert check method in a future update
     interestingGroups <- checkInterestingGroups(object, interestingGroups)
+
     object[["interestingGroups"]] <- NULL
     if (length(interestingGroups) > 1L) {
+        # For multiple groups, unite to a colon-separated string
         object <- unite(
             data = object,
             col = interestingGroups,
@@ -27,6 +44,7 @@ uniteInterestingGroups <- function(object, interestingGroups) {
     } else {
         object[["interestingGroups"]] <- object[[interestingGroups]]
     }
+
     # Set the `interestingGroups` column as factor
     object[["interestingGroups"]] <- as.factor(object[["interestingGroups"]])
     object