+++ tests/testthat/test_internal-prepareSampleMetadata.R
@@ -3,6 +3,10 @@ context("prepareSampleMetadata")
 test_that("Missing description column", {
     expect_error(
         .prepareSampleMetadata(mtcars),
-        "`description` column is required"
+        paste(
+            "is_subset :",
+            "The element 'description' in \"description\" is not in",
+            "colnames\\(object\\)."
+        )
     )
 })